Time: 0 ms (100%), Space: 8 MB (26.22%) - LeetHub

+++ 1009-complement-of-base-10-integer/1009-complement-of-base-10-integer.cpp
@@ -2,7 +2,7 @@ class Solution {
 public:
     int bitwiseComplement(int n) {
         if(n == 0) return 1;
-        long long x = pow(2, ((int)log2(n) + 1)) - 1;
+        int x = pow(2, ((int)log2(n) + 1)) - 1;
         return x ^ n;
     }
 };
